requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
coremltools==5.0
Flask==2.0.2
Flask-RESTful==0.3.9
flatbuffers==2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mpmath==1.2.1
numpy==1.21.2
onnx==1.10.1
onnxruntime==1.9.0
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sympy==1.9
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
